Refactor code and add Redis / Postgres connections
