+++ package.json
@@ -22,8 +22,10 @@
     "@codemirror/search": "^6.5.0",
     "@codemirror/state": "^6.2.1",
     "@codemirror/view": "^6.13.2",
+    "@redis/client": "^1.5.8",
     "express": "^4.18.2",
     "nanoid": "^4.0.2",
+    "postgres": "^3.3.5",
     "reconnecting-websocket": "^4.4.0",
     "remark-html": "^15.0.2",
     "remark-parse": "^10.0.2",
