+++ server/connection.ts
@@ -1,0 +1,79 @@
+import { Update } from "@codemirror/collab";
+import { ChangeSet, Text } from "@codemirror/state";
+import { WebSocket } from "ws";
+
+const initialText = `# CodeMirror Collaboration
+
+This is a collaborative **Markdown** editor shared by all viewers of this page. You just opened a new document, with your own personal link!
+
+Type here to see the output rendered below.
+
+Source code is available [here](https://github.com/ekzhang/cm-collab).`;
+
+class Document {
+  // The updates received so far (updates.length gives the current version)
+  updates: Update[] = [];
+
+  // The current document
+  doc = Text.of(initialText.split("\n"));
+
+  // Requests waiting for updates
+  pending: ((value: any) => void)[] = [];
+
+  handleMessage(ws: WebSocket, message: ArrayBuffer) {
+    const data = JSON.parse(message.toString());
+    console.log("Received:", data);
+
+    const resp = (value: any) => {
+      ws.send(JSON.stringify({ id: data.id, value }));
+    };
+
+    if (data.type == "pullUpdates") {
+      if (data.version < this.updates.length) {
+        resp(this.updates.slice(data.version));
+      } else {
+        this.pending.push(resp);
+      }
+    } else if (data.type == "pushUpdates") {
+      if (data.version != this.updates.length) {
+        resp(false);
+      } else {
+        for (let update of data.updates) {
+          // Convert the JSON representation to an actual ChangeSet instance
+          let changes = ChangeSet.fromJSON(update.changes);
+          this.updates.push({ changes, clientID: update.clientID });
+          this.doc = changes.apply(this.doc);
+        }
+        resp(true);
+        // Notify pending requests
+        while (this.pending.length) {
+          this.pending.pop()!(data.updates);
+        }
+      }
+    } else if (data.type == "getDocument") {
+      resp({ version: this.updates.length, doc: this.doc.toString() });
+    }
+  }
+}
+
+const documents: Map<string, Document> = new Map();
+
+export async function handleConnection(id: string, ws: WebSocket) {
+  console.log("New connection to document", id);
+
+  let doc = documents.get(id);
+  if (!doc) {
+    doc = new Document();
+    documents.set(id, doc);
+  }
+
+  ws.on("error", console.error);
+  ws.on("message", (message) => {
+    try {
+      doc!.handleMessage(ws, message as Buffer);
+    } catch (e: any) {
+      console.log("Failed to handle user message:", e.toString());
+      return;
+    }
+  });
+}
