+++ vite.config.ts
@@ -5,9 +5,10 @@ export default defineConfig({
     chunkSizeWarningLimit: 1000,
   },
   server: {
+    port: 7680,
     proxy: {
       "/ws": {
-        target: "ws://localhost:8080",
+        target: "ws://localhost:7671",
         changeOrigin: true,
         ws: true,
       },
